+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2025-01-24
+
+### Added
+
+- Fixed issue with unable to run sync in async context. Separated the sync and async prompt rendering into distinct classes.
+- Added support for async prompt rendering
+- Added tests for async prompt rendering
+- Added tests for previous issues
+
 ## [0.1.2] - 2025-01-24
 
 ### Added